+++ dvc/repo/experiments/branch.py
@@ -3,7 +3,7 @@ import logging
 from dvc.exceptions import InvalidArgumentError
 from dvc.repo import locked
 from dvc.repo.scm_context import scm_context
-from dvc.scm.base import RevError
+from dvc.scm import RevError
 
 from .base import InvalidExpRevError
 from .utils import exp_refs_by_rev
