+++ dvc/repo/experiments/ls.py
@@ -17,8 +17,7 @@ logger = logging.getLogger(__name__)
 @locked
 @scm_context
 def ls(repo, *args, rev=None, git_remote=None, all_=False, **kwargs):
-    from dvc.scm import resolve_rev
-    from dvc.scm.base import RevError
+    from dvc.scm import RevError, resolve_rev
     from dvc.scm.git import Git
 
     if rev:
